Version, env variables check, deployment script
