--- nodes/load_balancer/main.go
@@ -1,9 +1,0 @@
-package main
-
-import (
-	"bugout.dev/app-node-balancer/cmd"
-)
-
-func main() {
-	cmd.InitServer()
-}
