+++ nodes/node_balancer/configs/version.go
@@ -1,0 +1,3 @@
+package configs
+
+var NODE_BALANCER_VERSION = "0.0.1"
