+++ nodes/node_balancer/go.mod
@@ -1,0 +1,8 @@
+module github.com/bugout-dev/moonstream/nodes/node_balancer
+
+go 1.17
+
+require (
+	github.com/bugout-dev/humbug/go v0.0.0-20211206230955-57607cd2d205
+	github.com/google/uuid v1.3.0
+)
