+++ nodes/node_balancer/main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"github.com/bugout-dev/moonstream/nodes/node_balancer/cmd"
+)
+
+func main() {
+	cmd.InitServer()
+}
