+++ nodes/node_balancer/sample.env
@@ -1,0 +1,15 @@
+# Required environment variables for load balancer
+export MOONSTREAM_NODES_SERVER_PORT="<node_status_server_port>"
+export HUMBUG_REPORTER_NODE_BALANCER_TOKEN="<bugout_humbug_token_for_crash_reports>"
+
+# Ethereum nodes depends variables
+export MOONSTREAM_NODE_ETHEREUM_IPC_PORT="<node_geth_http_port>"
+
+export MOONSTREAM_NODE_ETHEREUM_A_IPC_ADDR="<node_geth_http_ip_addr>"
+export MOONSTREAM_NODE_ETHEREUM_B_IPC_ADDR="<node_geth_http_ip_addr>"
+
+# Polygon nodes depends variables
+export MOONSTREAM_NODE_POLYGON_IPC_PORT="<node_bor_http_port>"
+
+export MOONSTREAM_NODE_POLYGON_A_IPC_ADDR="<node_bor_http_ip_addr>"
+export MOONSTREAM_NODE_POLYGON_B_IPC_ADDR="<node_bor_http_ip_addr>"
